use flags of type double for cuda to limit deviations from oneAPI and Kokkos versions

+++ cuda/pagani/quad/GPUquad/Region_characteristics.cuh
@@ -11,7 +11,7 @@ class Region_characteristics {
 public:
   Region_characteristics(size_t num_regions)
   {
-    active_regions = quad::cuda_malloc<int>(num_regions * ndim);
+    active_regions = quad::cuda_malloc<double>(num_regions * ndim);
     sub_dividing_dim = quad::cuda_malloc<int>(num_regions * ndim);
     size = num_regions;
   }
@@ -23,7 +23,7 @@ public:
   }
 
   size_t size = 0;
-  int* active_regions = nullptr;
+  double* active_regions = nullptr;
   int* sub_dividing_dim = nullptr;
 };
